LayoutTests/imported/w3c/web-platform-tests/html/infrastructure/safe-passing-of-structured-data/shared-array-buffers/no-coop-coep.https.any-expected.txt


FAIL SharedArrayBuffer constructor does not exist without COOP+COEP assert_equals: expected (undefined) undefined but got (function) function "function SharedArrayBuffer() {
    [native code]
}"
PASS SharedArrayBuffer over MessageChannel without COOP+COEP
PASS SharedArrayBuffer over BroadcastChannel without COOP+COEP
FAIL SharedArrayBuffer over postMessage() without COOP+COEP assert_throws_dom: function "() => self.postMessage(sab)" did not throw
PASS Bonus: self.crossOriginIsolated
